+++ hutool-core/src/test/java/com/xiaoleilu/hutool/captcha/CaptchaTest.java
@@ -16,8 +16,9 @@ public class CaptchaTest {
 		
 		//定义图形验证码的长和宽
 		LineCaptcha lineCaptcha = CaptchaUtil.createLineCaptcha(200, 100);
+//		LineCaptcha lineCaptcha = new LineCaptcha(200, 100, 4, 150);
 		//图形验证码写出，可以写出到文件，也可以写出到流
-		lineCaptcha.write("d:/aaa.png");
+		lineCaptcha.write("d:/line.png");
 		//验证图形验证码的有效性，返回boolean值
 		lineCaptcha.verify("1234");
 	}
@@ -27,21 +28,23 @@ public class CaptchaTest {
 	public void circleCaptchaTest() {
 		
 		//定义图形验证码的长和宽
-		CircleCaptcha captcha = new CircleCaptcha(200, 100, 4, 20);
+		CircleCaptcha captcha = CaptchaUtil.createCircleCaptcha(200, 100, 4, 20);
+//		CircleCaptcha captcha = new CircleCaptcha(200, 100, 4, 20);
 		//图形验证码写出，可以写出到文件，也可以写出到流
-		captcha.write("d:/aaa.png");
+		captcha.write("d:/circle.png");
 		//验证图形验证码的有效性，返回boolean值
 		captcha.verify("1234");
 	}
 	
 	@Test
 	@Ignore
-	public void ShareCaptchaTest() {
+	public void ShearCaptchaTest() {
 		
 		//定义图形验证码的长和宽
-		ShearCaptcha captcha = new ShearCaptcha(200, 100, 4, 20);
+		ShearCaptcha captcha = CaptchaUtil.createShearCaptcha(200, 100, 4, 4);
+//		ShearCaptcha captcha = new ShearCaptcha(200, 100, 4, 4);
 		//图形验证码写出，可以写出到文件，也可以写出到流
-		captcha.write("d:/aaa.png");
+		captcha.write("d:/shear.png");
 		//验证图形验证码的有效性，返回boolean值
 		captcha.verify("1234");
 	}